SRC/relay/requirements.txt
websockets==11.0.3
kafka-python==2.0.2
asyncio==3.4.3
ujson==5.8.0